[FIX] VoucherService를 인터페이스로 추상화하라.
